pyproject.toml
[tool.poetry]
name = "pyerp"
version = "0.1.0"
description = "ERP"
authors = ["Avinash Gowda <avinashgowda.1994@gmail.com>"]
license = "BSD"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
django-nextjs = "^2.4.0"
djangorestframework = "^3.14.0"
markdown = "^3.5.2"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
